// FrontEnd/src/app/models/case.ts
export class Case {
    caseId: string = '';
    clientId: string = '';
    clientName: string = '';
    officerId: string = '';
    officerName: string = '';
    status: string = '';
    region: string = '';
    recommendations: string = '';
    createdDate: Date = new Date();
    updatedDate: Date = new Date();
}